// frontend/src/app/api/generated/model/customerDirectoryResponse.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CustomerDirectoryResponse {
  city: string | null;
  country: string | null;
  customer_number: string;
  email: string | null;
  first_name: string;
  id: number;
  is_active: boolean;
  last_name: string;
  phone: string | null;
}