// frontend/src/app/api/generated/model/customerProfileResponse.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CustomerProfileResponse {
  city: string | null;
  country: string | null;
  created_at: string;
  customer_number: string;
  date_of_birth: string | null;
  email: string | null;
  first_name: string;
  house_number: string | null;
  id: number;
  is_active: boolean;
  last_name: string;
  phone: string | null;
  postal_code: string | null;
  street: string | null;
  updated_at: string;
}