// frontend/src/app/api/generated/model/customerVerificationResponse.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { VerificationFactor } from './verificationFactor';

export interface CustomerVerificationResponse {
  customer_id: number;
  expires_at: string;
  factors: Array<VerificationFactor>;
  id: number;
  verified_at: string;
  verified_by_user_id: number;
}